no duplicate upvotes complete
